Refactor lyrics rendering
It is a bit simpler to understand and the code is split into multiple classes.
The main feature is that the lyrics are smooth scrolled without manually having the clear the top and bottom of the window. So there are no black bars above and below any more.
Another feature is that the scrolling is decoupled from the "lastPronouncedTick" so the yellow highlighting should be more accurate now.